+++ src/exports.c
@@ -2,11 +2,9 @@
 #include "Rinternals.h"
 
 SEXP strict_extract(SEXP e1, SEXP e2);
-SEXP unlock_env(SEXP env);
 
 static const R_CallMethodDef methods[] = {
     {"c_strict_extract", (DL_FUNC) &strict_extract, 2},
-    {"c_unlock_env", (DL_FUNC) &unlock_env, 1},
     {NULL, NULL, 0}
 };